+++ CHANGELOG.rst
@@ -6,6 +6,7 @@ Next version
 
 - Updated all dependencies.
 - Stopped putting anything into the global scope in ``init.js``.
+- Added optional support for showing typographic characters.
 
 
 0.4 (2024-05-26)